Update modeling_llama.py
The current implementation is wrong. This operation should interleave the values of the two vectors from the last dimension, rather than concatenate the two complete vectors. The former would lead to dim 2i and 2i+1 have the same value before applying cos and sin. The latter would lead to i and i + D/2 to have the same value before that.